MAINT split partial dependence and visualization (#16453)
